turn course schedule in two React components, delete schedule-html from index.html
