+++ public/index.html
@@ -17,65 +17,6 @@
       <!-- render of React-Component 'Home' -->
     </section>
 
-    <section id="schedule">
-      <h5>Kalender/Weekly Planner</h5>
-    <table>
-      <tr>
-        <th>Zeiten</th>
-        <th>Montag</th>
-        <th>Dienstag</th>
-        <th>Mittwoch</th>
-        <th>Donnerstag</th>
-        <th>Freitag</th>
-        <th>Samstag</th>
-      </tr>
-      <tr>
-        <td>8.00 - 9.30</td>
-        <td></td>
-        <td></td>
-        <td></td>
-        <td></td>
-        <td></td>
-        <td></td>
-      </tr>
-      <tr>
-        <td>9.45 - 11.15</td>
-        <td></td>
-        <td></td>
-        <td></td>
-        <td></td>
-        <td></td>
-        <td></td>
-      </tr>
-      <tr>
-        <td>12.15 - 13.45</td>
-        <td></td>
-        <td></td>
-        <td></td>
-        <td></td>
-        <td></td>
-        <td></td>
-      </tr>
-      <tr>
-        <td>14.00 - 15.30</td>
-        <td></td>
-        <td></td>
-        <td></td>
-        <td></td>
-        <td></td>
-        <td></td>
-      </tr>
-      <tr>
-        <td>15.45 - 17.15</td>
-        <td></td>
-        <td></td>
-        <td></td>
-        <td></td>
-        <td></td>
-        <td></td>
-      </tr>
-    </table>
-    </section>
    <footer>
       2016 &copy; Dein Semesterplaner
     </footer>
