+++ src/TimeSlots.js
@@ -1,0 +1,19 @@
+import React from 'react';
+
+class TimeSlots extends React.Component {
+  render() {
+    return (
+     <tr>
+      <td>8.30 - 9.45</td>
+      <td></td>
+      <td></td>
+      <td></td>
+      <td></td>
+      <td></td>
+      <td></td>
+    </tr>
+    )
+  }
+}
+
+export default TimeSlots
